Conflicts:
	apps/devApps/ofxQTKitVideoPlayer/ofxQTKitVideoPlayer.xcodeproj/jamesgeorge.pbxuser
	libs/openFrameworksCompiled/project/osx/openFrameworksLib.xcodeproj/jamesgeorge.perspectivev3

+++ addons/ofxiPhone/src/ofxiPhoneAppDelegate.h
@@ -40,7 +40,7 @@
 	NSTimer				*animationTimer;
     BOOL				animating;
     BOOL				displayLinkSupported;
-    NSInteger			animationFrameInterval;
+    float				animationFrameInterval;
     id					displayLink;
 	
 	EAGLView			*glView;
